requirements.txt
aniso8601==10.0.1
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
attrs==25.3.0
black==25.1.0
blinker==1.9.0
certifi==2025.4.26
click==8.2.0
distro==1.9.0
Flask==3.1.1
flask-restx==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mypy_extensions==1.1.0
openai==1.79.0
openapi==2.0.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
rpds-py==0.25.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
Werkzeug==3.1.3
